Django_Rest_API/travels/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
gunicorn
uvicorn
razorpay
python-dotenv